Contains all classes used for Zombies project
Actor - initialize the direction that the object is moving in (if
applicable), as well as all the default settings for the color.

Building - construct the components of a building in the city (with
default color and size), ensuring that the building is within the bounds
of the city.

City - the main class; draws the zombies, humans, and buildings onto the
grid, determines if the human is infected and therefore adds zombies,
and updates the state of the map at each time period.

Direction - an enumeration representing possible directions a human or
zombie can move.

DotPanel - consists of all the graphical components of the program
including the size of the pixel and color of the image.

Human - consists of all human capabilities. When a zombie is approaching
the relative area that a human is, the human starts to flee from the
zombie.

Zombie - consists of all zombie capabilities. Searches for a direction
in which a human can be found and starts to chase it in that direction.

ZombieSim - handles the overall game simulation with mouse clicks to
determine whether the game has been paused or is currently running.